Revert "Multi agent adjustment"
