+++ Pure/exceptions/StagnationException.py
@@ -8,6 +8,3 @@ class StagnationException(Exception):
     def __init__(self, message):
         self.message = message
         super().__init__(self.message)
-
-    def __str__(self):
-        return f"{self.message}"
